Refactor integration reporting to support comparison
